Remove DNS flaky packet injector code
This code does not work on Linux and works unreliably on macOS
for unspecified reasons. The existing code that provides a stub
resolver is already good enough, considering Docker. I think I
have elapsed all the time I could on this task for now, so, let's
surrended and call the the stub resolver good enough.