qa/telegram.py: QA telegram implementations
We can use this script to perform QA of telegram implementations. We are
using the command line interface of OONI v2.x because this is the same CLI
also used by Measurement Kit and by the miniooni client.

Currently, the miniooni client compiled so that we use the Go implementation
of telegram is already passing this suite of tests.