iptables: hijack ports 80 and 443 (#19)
* iptables: hijack ports 80 and 443

I have not figured out yet a way to hijack these two ports
for all users. The current solution is to assume that we are
running as root, and therefore root is whitelisted.

* Write integration testing for http/https hijack

Here the idea is basically to use sudo to run curl as nobody and make
sure we fail with a specific error code, to gain more confidence that we
are able to mess with http/https for non root.