Check all error conditions we care about w/ integration tests
This diff adds extra integration test checks that make sure we continue
to be able to reproduce the error conditions we care about.

The list of error conditions is such that we're able to produce all the
errors listed in the measure.sh[1] script by @fortuna.

Being able to produce all these errors should be a good enough starting
point for evaluating different web measurement strategies.

.. [1] https://github.com/Jigsaw-Code/net-analysis/blob/master/netanalysis/blocktest/measure.sh

Closes #7