Allow to run tests on non Linux systems
Of course, the coverage is gonna be lower. We don't specify with tests
the behaviour on non Linux systems. The current codebase will panic when
trying to apply an iptables censoring policy. I like this behaviour
because it's better than running with a warning without iptables, which
means we're not doing what we are actually supposed to do.